Refactored query services
