+++ samples/event-sourcing-esdb-simple/src/main/java/io/eventdriven/ecommerce/shoppingcarts/ShoppingCartsConfig.java
@@ -34,16 +34,5 @@ class ShoppingCartsConfig {
       ShoppingCart::empty
     );
   }
-
-  @Bean
-  ShoppingCartQueryService shoppingCartQueryService(
-    ShoppingCartDetailsRepository detailsRepository,
-    ShoppingCartShortInfoRepository shortInfoRepository
-  ) {
-    return new ShoppingCartQueryService(
-      detailsRepository,
-      shortInfoRepository
-    );
-  }
 }
