+++ samples/talk/src/test/java/io/eventdriven/shoppingcarts/ShoppingCartCommand.java
@@ -1,0 +1,10 @@
+package io.eventdriven.shoppingcarts;
+
+import io.eventdriven.shoppingcarts.productitems.PricedProductItem;
+import io.eventdriven.shoppingcarts.productitems.ProductItem;
+
+import java.time.OffsetDateTime;
+import java.util.UUID;
+
+public interface ShoppingCartCommand {
+}
