+++ samples/talk/src/test/java/io/eventdriven/shoppingcarts/entities/CommandHandler.java
@@ -1,0 +1,15 @@
+package io.eventdriven.shoppingcarts.entities;
+
+import com.eventstore.dbclient.*;
+import io.eventdriven.shoppingcarts.serialization.EventSerializer;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.function.BiFunction;
+import java.util.function.Function;
+import java.util.function.Supplier;
+
+public class CommandHandler {
+
+}
